+++ src/test/java/org/ballerinalang/completion/BallerinaHiddenTemplateCompletionTest.java
@@ -1,0 +1,115 @@
+/*
+ *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.ballerinalang.completion;
+
+import java.util.LinkedList;
+import java.util.List;
+
+public class BallerinaHiddenTemplateCompletionTest extends BallerinaCompletionTestBase {
+
+    /**
+     * Hidden templates.
+     */
+    public void testPackageKeyword() {
+        doCheckResult("test.bal", "package<caret>", "package ", null);
+    }
+
+    public void testImportKeyword() {
+        doCheckResult("test.bal", "import<caret>", "import ", null);
+    }
+
+    public void testIfKeyword() {
+        doCheckResult("test.bal", "function test(){ if<caret> }", "function test(){ if () {\n    \n} }", null);
+    }
+
+    public void testElseKeyword() {
+        doCheckResult("test.bal", "function test(){ else<caret> }", "function test(){ else {\n    \n} }", null);
+    }
+
+    public void testForkKeyword() {
+        doCheckResult("test.bal", "function test(){ fork<caret> }", "function test(){ fork {\n    \n} }", null);
+    }
+
+    public void testJoinKeyword() {
+        doCheckResult("test.bal", "function test(){ join<caret> }", "function test(){ join () (message[] ) {\n    \n}" +
+                " }", null);
+    }
+
+    public void testTimeoutKeyword() {
+        doCheckResult("test.bal", "function test(){ timeout<caret> }", "function test(){ timeout () (message[] ) " +
+                "{\n    \n} }", null);
+    }
+
+    public void testWorkerKeyword() {
+        doCheckResult("test.bal", "function test(){ worker<caret> }", "function test(){ worker  {\n    \n} }",
+                null);
+    }
+
+    public void testTransformKeyword() {
+        doCheckResult("test.bal", "function test(){ transform<caret> }", "function test(){ transform {\n    \n} }",
+                null);
+    }
+
+    public void testTransactionKeyword() {
+        doCheckResult("test.bal", "function test(){ transaction<caret> }", "function test(){ transaction {\n    \n} " +
+                "}", null);
+    }
+
+    public void testAbortedKeyword() {
+        doCheckResult("test.bal", "function test(){ aborted<caret> }", "function test(){ aborted {\n    \n} }", null);
+    }
+
+    public void testCommittedKeyword() {
+        doCheckResult("test.bal", "function test(){ committed<caret> }", "function test(){ committed {\n    \n} }",
+                null);
+    }
+
+    public void testTryKeyword() {
+        doCheckResult("test.bal", "function test(){ try<caret> }", "function test(){ try {\n    \n} }", null);
+    }
+
+    public void testCatchKeyword() {
+        doCheckResult("test.bal", "function test(){ catch<caret> }", "function test(){ catch () {\n    \n} }",
+                null);
+    }
+
+    public void testFinallyKeyword() {
+        doCheckResult("test.bal", "function test(){ finally<caret> }", "function test(){ finally {\n    \n} }", null);
+    }
+
+    public void testIterateKeyword() {
+        doCheckResult("test.bal", "function test(){ iterate<caret> }", "function test(){ iterate ( : ) {\n    \n} }",
+                null);
+    }
+
+    public void testWhileKeyword() {
+        doCheckResult("test.bal", "function test(){ while<caret> }", "function test(){ while () {\n    \n} }",
+                null);
+    }
+
+    public void testContinueKeyword() {
+        doCheckResult("test.bal", "function test(){ continue<caret> }", "function test(){ continue; }", null);
+    }
+
+    public void testBreakKeyword() {
+        doCheckResult("test.bal", "function test(){ break<caret> }", "function test(){ break; }", null);
+    }
+
+    public void testThrowKeyword() {
+        doCheckResult("test.bal", "function test(){ throw<caret> }", "function test(){ throw ; }", null);
+    }
+}
